+++ crates/rpc/src/server/api.rs
@@ -2,13 +2,7 @@ use std::sync::Arc;
 use std::time::Duration;
 
 use anyhow::Context;
-use miden_node_proto::clients::{
-    BlockProducer,
-    BlockProducerClient,
-    Builder,
-    StoreRpc,
-    StoreRpcClient,
-};
+use miden_node_proto::clients::{BlockProducerClient, Builder, StoreRpcClient};
 use miden_node_proto::errors::ConversionError;
 use miden_node_proto::generated::rpc::api_server::{self, Api};
 use miden_node_proto::generated::{self as proto};
@@ -59,7 +53,8 @@ impl RpcService {
                 .without_timeout()
                 .without_metadata_version()
                 .without_metadata_genesis()
-                .connect_lazy::<StoreRpc>()
+                .with_otel_context_injection()
+                .connect_lazy::<StoreRpcClient>()
         };
 
         let block_producer = block_producer_url.map(|block_producer_url| {
@@ -73,7 +68,8 @@ impl RpcService {
                 .without_timeout()
                 .without_metadata_version()
                 .without_metadata_genesis()
-                .connect_lazy::<BlockProducer>()
+                .with_otel_context_injection()
+                .connect_lazy::<BlockProducerClient>()
         });
 
         Self {